+++ PHP/CGA/project/group/group-solution.php
@@ -1,15 +1,50 @@
+<script>
+    function valdiateSolution() {
+
+        var solution_content, file;
+
+        solution_content = document.getElementById("solution_content").value;
+        file = document.getElementById("file").value;
+
+        if (solution_content == '') {
+            alert("Please enter a solution content.");
+            document.getElementById("solution_content").focus();
+            return false;
+        } else if (file == '') {
+            alert("Please enter file.");
+            document.getElementById("file").focus();
+            return false;
+        } else
+            return true;
+    }
+
+    function validateGrade() {
+
+        var grade;
+
+        grade = document.getElementById("grade").value;
+
+        if (grade == '') {
+            alert("Please enter a grade.");
+            document.getElementById("grade").focus();
+            return false;
+        } else
+            return true;
+    }
+</script>
 <?php
 
-$user_id = $_SESSION['user_id'];
+$session_user_id = $_SESSION['user_id'];
 
 if (isset($_GET['course_id'])) {
-    $course_id = $_GET['course_id'];
+    $session_course_id = $_GET['course_id'];
+}
+if (isset($_GET['group_id'])) {
+    $session_group_id = $_GET['group_id'];
 }
 
 // ADD
 if (isset($_POST['upload_file'])) {
-
-    pre_print($_REQUEST);
 
     $task_id = $_GET['task_id'];
 
@@ -34,7 +69,7 @@ if (isset($_POST['upload_file'])) {
                 VALUES('$solution_type', '$solution_content', '$task_id', '$file_id')";
 
         if (mysqli_query($conn, $add)) {
-            array_push($success, "solution added Successful");
+            array_push($success, "Solution added Successful");
         } else {
             array_push($errors, "Error adding solution: " . mysqli_error($conn));
         }
@@ -84,16 +119,53 @@ if (isset($_POST['update_file'])) {
 // DELETE
 if (isset($_GET['delete_id'])) {
     $id = mysqli_real_escape_string($conn, $_GET['delete_id']);
-
-    $file_id = $_GET['delete_file'];
-
     $delete = "DELETE FROM solution WHERE solution_id='$id'";
     if (mysqli_query($conn, $delete)) {
-        delete_file($file_id);
+        delete_file($_GET['delete_file']);
         array_push($success, "Delete successful");
+        header("location: ?page=group-solution&course_id=$session_course_id");
     } else {
         array_push($errors, "Delete error: " . mysqli_error($conn));
     }
+}
+
+// ADD GRADE
+if (isset($_POST['add_grade'])) {
+
+    $group_id = $_GET['group_id'];
+    $solution_id = $_GET['solution_id'];
+
+    // receive all input values from the form
+    $grade = mysqli_real_escape_string($conn, $_POST['grade']);
+
+    // form validation: ensure that the form is correctly filled ...
+    // by adding (array_push()) corresponding error unto $errors array
+    if (empty($grade)) {
+        array_push($errors, "Grade is required");
+    } elseif ($grade > 100 || $grade < 0) {
+        array_push($errors, "Invalid grade");
+    }
+
+    $query = "SELECT * FROM student_groups as g
+                JOIN member_of_group as mg ON mg.group_id = g.group_id
+                    JOIN student as st ON st.student_id = mg.student_id
+                    WHERE g.group_id = '$group_id'";
+
+    $groupArr = mysqli_query($conn, $query);
+
+    foreach ($groupArr as $row) {
+        $student_id = $row['student_id'];
+
+        if (count($errors) == 0) {
+            $add_grade = "INSERT INTO grades (grade, student_id, solution_id)
+                            VALUES('$grade', '$student_id', '$solution_id')";
+            if (!mysqli_query($conn, $add_grade)) {
+                array_push($errors, "Error adding grade: " . mysqli_error($conn));
+            }
+        }
+    }
+
+    array_push($success, "Grade added Successful");
 }
 
 ?>
@@ -104,26 +176,47 @@ if (isset($_GET['delete_id'])) {
     display_success();
     display_error();
 
-    $query = "SELECT t.*, c.*, f.*, s.solution_id, s.solution_type, s.solution_content, u.* FROM task as t
-    JOIN course as c ON c.course_id = t.course_id
-	JOIN user_course_section as ucs ON ucs.course_id = c.course_id
-	JOIN users as us ON us.user_id = ucs.user_id
-    LEFT JOIN solution as s ON s.task_id = t.task_id
-	LEFT JOIN files as f ON f.file_id = s.file_id
-    LEFT JOIN users as u ON u.user_id = f.uploaded_by_uid
-    WHERE us.user_id = '$user_id' AND c.course_id = '$course_id'
-    ORDER BY t.task_id ASC";
+    if (isStudent()) {
+        $query = "SELECT t.*, c.*, f.*, s.solution_id, s.solution_type, s.solution_content, u.*, g.* FROM task as t
+        JOIN course as c ON c.course_id = t.course_id
+        JOIN group_of_course as gc ON gc.course_id = c.course_id
+        JOIN student_groups as g ON g.group_id = gc.group_id
+        JOIN user_course_section as ucs ON ucs.course_id = c.course_id
+        JOIN users as us ON us.user_id = ucs.user_id
+        LEFT JOIN solution as s ON s.task_id = t.task_id
+        LEFT JOIN files as f ON f.file_id = s.file_id
+        LEFT JOIN users as u ON u.user_id = f.uploaded_by_uid
+        WHERE us.user_id = '$session_user_id' AND c.course_id = '$session_course_id' AND g.group_id = '$session_group_id'
+        ORDER BY t.task_id ASC";
+    } else {
+        $query = "SELECT t.*, c.*, f.*, s.solution_id, s.solution_type, s.solution_content, u.* FROM task as t
+        JOIN course as c ON c.course_id = t.course_id
+        JOIN user_course_section as ucs ON ucs.course_id = c.course_id
+        JOIN users as us ON us.user_id = ucs.user_id
+        LEFT JOIN solution as s ON s.task_id = t.task_id
+        LEFT JOIN files as f ON f.file_id = s.file_id
+        LEFT JOIN users as u ON u.user_id = f.uploaded_by_uid
+        WHERE us.user_id = '$session_user_id' AND c.course_id = '$session_course_id'
+        ORDER BY t.task_id ASC";
+    }
     $results = mysqli_query($conn, $query);
 
-    $course_name = mysqli_fetch_assoc($results)['course_name'];
-
+    if (mysqli_num_rows($results) > 0) {
+        $course_name = mysqli_fetch_assoc($results)['course_name'];
+    } else {
+        $course_name = "No";
+    }
     ?>
+
     <h2><?= $course_name ?> Solutions</h2>
     <hr>
     <table>
         <thead>
             <tr>
                 <th>Task</th>
+                <?php if (isStudent()) { ?>
+                    <th>Group Name</th>
+                <?php } ?>
                 <th>Type</th>
                 <th>Content</th>
                 <th>Uploaded by</th>
@@ -132,6 +225,8 @@ if (isset($_GET['delete_id'])) {
                 <?php
                 if (isStudent()) {
                     echo '<th colspan="4">Action</th>';
+                } else if (isProfessor()) {
+                    echo '<th colspan="2">Action</th>';
                 } else {
                     echo '<th>Action</th>';
                 }
@@ -141,40 +236,75 @@ if (isset($_GET['delete_id'])) {
         <tbody>
             <?php
             foreach ($results as $row) {
+
                 $task_id = $row['task_id'];
+                $task_content = $row['task_content'];
+                $task_deadline = $row['task_deadline'];
+                $today = date('Y-m-d', time());
 
                 $solution_type = $row['solution_type'];
                 $solution_id = $row['solution_id'];
                 $solution_content = $row['solution_content'];
-                $uploaded_by_uid = $row['username'];
+
+                $uploaded_by_uid = $row['uploaded_by_uid'];
+                $uploaded_by = $row['username'];
+
                 if ($row['uploaded_on'] !== NULL) {
                     $uploaded_on = date_convert($row['uploaded_on']);
+                    if (isStudent()) {
+                        $group_id = $row['group_id'];
+                        $group_name = $row['group_name'];
+                    }
                 } else {
-                    $uploaded_on = '';
+                    $uploaded_on = $group_id = $group_name = "";
                 }
+
+                if (isStudent()) {
+                    $group_leader_sid = $row['group_leader_sid'];
+                }
+
                 $file_id = $row['file_id'];
                 $file_name = $row['file_name'];
-                $task_content = $row['task_content'];
                 $course_id = $row['course_id'];
             ?>
                 <tr>
                     <td><b><a href='?page=course-task&course_id=<?= $course_id ?>'><?= $task_content ?></a></b></td>
+                    <?php if (isStudent()) { ?>
+                        <td><b><?= $group_name ?></b></td>
+                    <?php } ?>
                     <td><?= $solution_type ?></td>
                     <td><?= $solution_content ?></td>
-                    <td><?= $uploaded_by_uid ?></td>
+                    <td><?= $uploaded_by ?></td>
                     <td><?= $uploaded_on ?></td>
                     <td><?= $file_name ?></td>
+
                     <?php
                     if (isStudent()) {
-                        if ($file_id == NULL && $solution_id == NULL) {
-                            echo "<td><a href='?page=group-solution&course_id=$course_id&task_id=$task_id&upload_view=true'>Upload</a></td>";
+                        $session_student_id = mysqli_fetch_assoc(get_records_where('student', 'user_id', $session_user_id))['student_id'];
+                        if (isGroupLeader($session_student_id, $session_group_id)) {
+                            if ($file_id == NULL && $solution_id == NULL) {
+                                if ($task_deadline >= $today) {
+                                    echo "<td><a href='?page=group-solution&course_id=$course_id&task_id=$task_id&group_id=$session_group_id&upload_view=true'>Upload</a></td>";
+                                } else {
+                                    echo "<td><b style='color:red;'>Deadline passed</b></td>";
+                                }
+                            } else {
+                                echo "<td><a href='?page=group-solution&course_id=$course_id&download_file=$file_id'>Download</a></td>";
+                                echo "<td><a href='?page=group-solution&course_id=$course_id&update_id=$solution_id&update_file=$file_id&update_view=true&group_id=$group_id'>Update</a></td>";
+                                echo "<td><a href='?page=group-solution&course_id=$course_id&delete_id=$solution_id&delete_file=$file_id' onclick='return confirm(&quot;Are you sure you want to delete?&quot;)'>Delete</a></td>";
+                            }
                         } else {
-                            echo "<td><a href='?page=group-solution&course_id=$course_id&download_file=$file_id'>Download</a></td>";
-                            echo "<td><a href='?page=group-solution&course_id=$course_id&update_id=$solution_id&update_file=$file_id&update_view=true'>Update</a></td>";
-                            echo "<td><a href='?page=group-solution&course_id=$course_id&delete_id=$solution_id&delete_file=$file_id' onclick='return confirm(&quot;Are you sure you want to delete?&quot;)'>Delete</a></td>";
+                            if ($file_id == NULL && $solution_id == NULL) {
+                                echo "<td>No Solution</td>";
+                            } else {
+                                echo "<td><a href='?page=group-solution&course_id=$course_id&download_file=$file_id'>Download</a></td>";
+                            }
                         }
                     } elseif ($file_id !== NULL) {
                         echo "<td><a href='?page=group-solution&course_id=$course_id&download_file=$file_id'>Download</a></td>";
+                        if (isProfessor()) {
+                            echo "<td><a href='?page=group-solution&course_id=$course_id&group_id=$group_id&solution_id=$solution_id&grade_view=true'>Grade</a></td>";
+                        }
                     } else {
                         echo "<td>No Solution</td>";
                     }
@@ -202,7 +332,7 @@ if (isset($_GET['delete_id'])) {
 
             <hr>
             <div class="form-container">
-                <form class="form-body" action="" enctype="multipart/form-data" method="POST">
+                <form class="form-body" action="" enctype="multipart/form-data" method="POST" onsubmit="return valdiateSolution()">
 
                     <h3>Upload solution</h3>
 
@@ -212,7 +342,7 @@ if (isset($_GET['delete_id'])) {
                     </div>
 
                     <div class="form-input">
-                        <label for="solution_type">solution type</label>
+                        <label for="solution_type">Solution type</label>
                         <span>
                             <select name="solution_type">
                                 <option value="<?= $task_type ?>" selected><?= $task_type ?></option>
@@ -222,12 +352,12 @@ if (isset($_GET['delete_id'])) {
 
                     <div class="form-input">
                         <label>Description</label>
-                        <span><input type="text" name="solution_content"></span>
+                        <span><input type="text" name="solution_content" id="solution_content"></span>
                     </div>
 
                     <div class="form-input">
                         <label>Select file</label>
-                        <span><input type="file" name="file"></span>
+                        <span><input type="file" name="file" id="file"></span>
                     </div>
 
                     <div class="form-submit">
@@ -256,7 +386,7 @@ if (isset($_GET['delete_id'])) {
 
             <hr>
             <div class="form-container">
-                <form class="form-body" action="" enctype="multipart/form-data" method="POST">
+                <form class="form-body" action="" enctype="multipart/form-data" method="POST" onsubmit="return valdiateSolution()">
 
                     <h3>Update File</h3>
 
@@ -271,17 +401,63 @@ if (isset($_GET['delete_id'])) {
 
                     <div class="form-input">
                         <label>Description</label>
-                        <span><input type="text" name="solution_content" value="<?= $solution_content ?>"></span>
+                        <span><input type="text" name="solution_content" id="solution_content" value="<?= $solution_content ?>"></span>
                     </div>
 
 
                     <div class=" form-input">
                         <label>Select file</label>
-                        <span><input type="file" name="file"> </span>
+                        <span><input type="file" name="file" id="file"> </span>
                     </div>
 
                     <div class="form-submit">
                         <input type="submit" name="update_file" value="Update">
+                    </div>
+
+                </form>
+            </div>
+
+        <?php } ?>
+
+    <?php } ?>
+
+    <?php if (isProfessor()) { ?>
+
+        <?php if (isset($_GET['grade_view'])) { ?>
+
+            <?php
+
+            $group_id = $_GET['group_id'];
+            $group_name = mysqli_fetch_assoc(get_records_where('student_groups', 'group_id', $group_id))['group_name'];
+
+            $solution_id = $_GET['solution_id'];
+            $solution_content = mysqli_fetch_assoc(get_records_where('solution', 'solution_id', $solution_id))['solution_content'];
+
+            ?>
+
+            <hr>
+            <div class="form-container">
+                <form class="form-body" action="" enctype="multipart/form-data" method="POST" onsubmit="return validateGrade()">
+
+                    <h3>Grade solution</h3>
+
+                    <div class="form-input">
+                        <label>Group</label>
+                        <span><b><?= $group_name ?></b></span>
+                    </div>
+
+                    <div class="form-input">
+                        <label>Solution for</label>
+                        <span><b><?= $solution_content ?></b></span>
+                    </div>
+
+                    <div class="form-input">
+                        <label>Grade</label>
+                        <span><input type="number" name="grade" id="grade"></span>
+                    </div>
+
+                    <div class="form-submit">
+                        <input type="submit" name="add_grade" value="Add Grade">
                     </div>
 
                 </form>
